attempt to improve signup activity

+++ app/src/main/java/com/example/restauranttracker/MainActivity.java
@@ -194,7 +194,7 @@ public class MainActivity extends AppCompatActivity {
         startActivity(LoginActivity.loginIntentFactory(getApplicationContext()));
     }
 
-    static Intent mainActivityIntentFactory(Context context, int userId) {
+    static Intent mainActivityIntentFactory(Context context, long userId) {
         Intent intent = new Intent(context, MainActivity.class);
         intent.putExtra(MAIN_ACTIVITY_USER_ID, userId);
         return intent;
